feat: centralize auth logic on auth guard
