+++ apps/web-shell/pages/_app.tsx
@@ -2,6 +2,7 @@ import type { AppProps } from 'next/app';
 import '@fiap-farms/web-ui/global.css';
 import { AppCacheProvider } from '@mui/material-nextjs/v14-pagesRouter';
 import { useAuthListener } from '@fiap-farms/auth-store';
+import { AuthGuard } from '../components/AuthGuard';
 
 export default function App({ Component, pageProps }: AppProps) {
   // Setup Firebase auth state listener at app level
@@ -9,7 +10,9 @@ export default function App({ Component, pageProps }: AppProps) {
 
   return (
     <AppCacheProvider {...pageProps}>
-      <Component {...pageProps} />
+      <AuthGuard>
+        <Component {...pageProps} />
+      </AuthGuard>
     </AppCacheProvider>
   );
 }
